Docs

+++ shift/mpiutils/mpiclass.py
@@ -12,21 +12,19 @@ class MPI:
         Initialises MPI.
         """
         from mpi4py import MPI as mpi
-        
+
         self.mpi = mpi
         self.comm = mpi.COMM_WORLD
         self.rank = self.comm.Get_rank()
         self.size = self.comm.Get_size()
         self.loop_size = None
-        self.mpi_info = 'Proc ' + str(self.rank+1)+' of ' + str(self.size)
-
+        self.mpi_info = "Proc " + str(self.rank + 1) + " of " + str(self.size)
 
     def wait(self):
         """
         Tells all jobs to wait -- to ensure jobs are synchronised.
         """
         self.comm.Barrier()
-
 
     def set_loop(self, loop_size: int) -> int:
         """
@@ -43,7 +41,6 @@ class MPI:
         self.loop_size = loop_size
         return loops.get_MPI_loop_size(loop_size, self.size)
 
-
     def mpi_ind2ind(self, mpi_ind: int) -> int:
         """
         Converts the MPI_ind of a distributed loop to the index of a full loop.
@@ -58,15 +55,15 @@ class MPI:
         """
         return loops.MPI_ind2ind(mpi_ind, self.rank, self.size, self.loop_size)
 
-
     def clean_loop(self):
         """
         Gets ride of loop_size definition.
         """
         self.loop_size = None
 
-
-    def split(self, length: int, size: Optional[int]=None) -> Tuple[np.ndarray, np.ndarray]:
+    def split(
+        self, length: int, size: Optional[int] = None
+    ) -> Tuple[np.ndarray, np.ndarray]:
         """
         For splitting an array across nodes.
 
@@ -77,7 +74,7 @@ class MPI:
         size: int
             Size of the MPI.size (i.e. MPI task), if set then this will be used to be split
             for other reasons.
-        
+
         Returns
         -------
         split1 : array_like
@@ -86,27 +83,26 @@ class MPI:
             The indices of the last element of the split array.
         """
         if size is None:
-            split_equal = length/self.size
+            split_equal = length / self.size
         else:
-            split_equal = length/size
+            split_equal = length / size
         split_floor = np.floor(split_equal)
         split_remain = split_equal - split_floor
         if size is None:
-            counts = split_floor*np.ones(self.size)
-            counts[:int(np.round(split_remain*self.size, decimals=0))] += 1
+            counts = split_floor * np.ones(self.size)
+            counts[: int(np.round(split_remain * self.size, decimals=0))] += 1
         else:
-            counts = split_floor*np.ones(size)
-            counts[:int(np.round(split_remain*size, decimals=0))] += 1
-        counts = counts.astype('int')
+            counts = split_floor * np.ones(size)
+            counts[: int(np.round(split_remain * size, decimals=0))] += 1
+        counts = counts.astype("int")
         if size is None:
-            splits = np.zeros(self.size+1, dtype='int')
+            splits = np.zeros(self.size + 1, dtype="int")
         else:
-            splits = np.zeros(size+1, dtype='int')
+            splits = np.zeros(size + 1, dtype="int")
         splits[1:] = np.cumsum(counts)
         split1 = splits[:-1]
         split2 = splits[1:]
         return split1, split2
-
 
     def split_array(self, array: np.ndarray) -> np.ndarray:
         """
@@ -122,10 +118,11 @@ class MPI:
         Split array.
         """
         split1, split2 = self.split(len(array))
-        return array[split1[self.rank]:split2[self.rank]]
+        return array[split1[self.rank] : split2[self.rank]]
 
-
-    def check_partition(self, NDshape: List[int], NDshape_split: List[int]) -> np.ndarray:
+    def check_partition(
+        self, NDshape: List[int], NDshape_split: List[int]
+    ) -> np.ndarray:
         """
         Returns a boolean array showing the axes that an array will be split along.
 
@@ -142,9 +139,8 @@ class MPI:
         """
         return np.array([NDshape[i] == NDshape_split[i] for i in range(len(NDshape))])
 
-
     # TODO: Remove these two functions -- pretty certain these are defunct and no longer used elsewhere.
-    
+
     # def create_split_ndarray(self, arrays_nd: np.ndarray, whichaxis: List[bool]) -> np.ndarray:
     #     """
     #     Splits a list of 1D arrays based on the data partitioning scheme. To be used with
@@ -156,7 +152,7 @@ class MPI:
     #         List of 1D arrays to be split.
     #     whichaxis : array_like
     #         Boolean array showing whether array will not be split along a said axes.
-        
+
     #     Returns
     #     -------
     #     split_arrays : array_like
@@ -172,7 +168,6 @@ class MPI:
     #             split_arrays.append(_array)
     #     return split_arrays
 
-
     # def create_split_ndgrid(self, arrays_nd: np.ndarray, whichaxis: List[bool]) -> np.ndarray:
     #     """
     #     Creates a partitioned gridded data set.
@@ -183,7 +178,7 @@ class MPI:
     #         List of arrays to be split.
     #     whichaxis : array_like
     #         Boolean array showing whether array will not be split along a said axes.
-        
+
     #     Returns
     #     -------
     #     split_grid : array_like
@@ -193,14 +188,12 @@ class MPI:
     #     split_grid = np.meshgrid(*split_arrays, indexing='ij')
     #     return split_grid
 
-
     def mpi_print(self, *value: Any) -> None:
         """
         Python print function using flush=True so print statements are outputed
         immediately in an MPI setting.
         """
         print(*value, flush=True)
-
 
     def mpi_print_zero(self, *value: Any) -> None:
         """
@@ -209,8 +202,9 @@ class MPI:
         if self.rank == 0:
             self.mpi_print(*value)
 
-
-    def send(self, data: np.ndarray, to_rank: Optional[int]=None, tag: int=11) -> None:
+    def send(
+        self, data: np.ndarray, to_rank: Optional[int] = None, tag: int = 11
+    ) -> None:
         """
         Sends data from current core to other specified or all cores.
 
@@ -230,8 +224,7 @@ class MPI:
                 if i != self.rank:
                     self.comm.send(data, dest=i, tag=tag)
 
-
-    def recv(self, from_rank: int, tag: int=11) -> np.ndarray:
+    def recv(self, from_rank: int, tag: int = 11) -> np.ndarray:
         """
         Receive data from another node.
 
@@ -250,7 +243,6 @@ class MPI:
         data = self.comm.recv(source=from_rank, tag=tag)
         return data
 
-
     def broadcast(self, data: Any) -> Any:
         """
         Broadcast data from rank=0 to all nodes.
@@ -262,30 +254,28 @@ class MPI:
         self.wait()
         return data
 
-
     def send_up(self, data: Any) -> Any:  # pragma: no cover
         """
         Send data from each node to the node above.
         """
-        
+
         datain = np.copy(data)
 
-        if self.rank < self.size-1:
-            self.send(datain, to_rank=self.rank+1, tag=10+self.rank)
+        if self.rank < self.size - 1:
+            self.send(datain, to_rank=self.rank + 1, tag=10 + self.rank)
         if self.rank > 0:
-            dataout = self.recv(self.rank-1, tag=10+self.rank-1)
-        
+            dataout = self.recv(self.rank - 1, tag=10 + self.rank - 1)
+
         self.wait()
 
-        if self.rank == self.size-1:
-            self.send(datain, to_rank=0, tag=10+self.size)
+        if self.rank == self.size - 1:
+            self.send(datain, to_rank=0, tag=10 + self.size)
         if self.rank == 0:
-            dataout = self.recv(self.size-1, tag=10+self.size)
-        
+            dataout = self.recv(self.size - 1, tag=10 + self.size)
+
         self.wait()
 
         return dataout
-
 
     def isend_up(self, data: Any) -> Any:
         """
@@ -295,21 +285,22 @@ class MPI:
 
         # Non-blocking receive from below
         if self.rank > 0:
-            req_recv = self.comm.irecv(source=self.rank-1, tag=10+self.rank-1)
+            req_recv = self.comm.irecv(source=self.rank - 1, tag=10 + self.rank - 1)
         else:
-            req_recv = self.comm.irecv(source=self.size-1, tag=10+self.size)
+            req_recv = self.comm.irecv(source=self.size - 1, tag=10 + self.size)
 
         # Non-blocking send upward
-        if self.rank < self.size-1:
-            req_send = self.comm.isend(obj=datain, dest=self.rank+1, tag=10+self.rank)
+        if self.rank < self.size - 1:
+            req_send = self.comm.isend(
+                obj=datain, dest=self.rank + 1, tag=10 + self.rank
+            )
         else:
-            req_send = self.comm.isend(obj=datain, dest=0, tag=10+self.size)
+            req_send = self.comm.isend(obj=datain, dest=0, tag=10 + self.size)
 
         # Wait for receive and send to complete
         dataout = req_recv.wait()
         req_send.wait()
         return dataout
-    
 
     def send_down(self, data: Any) -> Any:  # pragma: no cover
         """
@@ -318,21 +309,20 @@ class MPI:
         datain = np.copy(data)
 
         if self.rank > 0:
-            self.send(datain, to_rank=self.rank-1, tag=20+self.rank)
-        if self.rank < self.size-1:
-            dataout = self.recv(self.rank+1, tag=20+self.rank+1)
-        
+            self.send(datain, to_rank=self.rank - 1, tag=20 + self.rank)
+        if self.rank < self.size - 1:
+            dataout = self.recv(self.rank + 1, tag=20 + self.rank + 1)
+
         self.wait()
 
-        if self.rank == self.size-1:
-            dataout = self.recv(0, tag=20+self.size)
+        if self.rank == self.size - 1:
+            dataout = self.recv(0, tag=20 + self.size)
         if self.rank == 0:
-            self.send(datain, to_rank=self.size-1, tag=20+self.size)
-        
+            self.send(datain, to_rank=self.size - 1, tag=20 + self.size)
+
         self.wait()
 
         return dataout
-
 
     def isend_down(self, data: Any) -> Any:
         """
@@ -341,24 +331,27 @@ class MPI:
         datain = np.copy(data)
 
         # Non-blocking receive from above
-        if self.rank < self.size-1:
-            req_recv = self.comm.irecv(source=self.rank+1, tag=20+self.rank+1)
+        if self.rank < self.size - 1:
+            req_recv = self.comm.irecv(source=self.rank + 1, tag=20 + self.rank + 1)
         else:
-            req_recv = self.comm.irecv(source=0, tag=20+self.size)
+            req_recv = self.comm.irecv(source=0, tag=20 + self.size)
 
         # Non-blocking send downward
         if self.rank > 0:
-            req_send = self.comm.isend(obj=datain, dest=self.rank-1, tag=20+self.rank)
+            req_send = self.comm.isend(
+                obj=datain, dest=self.rank - 1, tag=20 + self.rank
+            )
         else:
-            req_send = self.comm.isend(obj=datain, dest=self.size-1, tag=20+self.size)
+            req_send = self.comm.isend(
+                obj=datain, dest=self.size - 1, tag=20 + self.size
+            )
 
         # Wait for receive and send to complete
         dataout = req_recv.wait()
         req_send.wait()
         return dataout
 
-
-    def collect(self, data: np.ndarray, outlist: bool=False) -> np.ndarray:
+    def collect(self, data: np.ndarray, outlist: bool = False) -> np.ndarray:
         """
         Collects a distributed data to the processor with rank=0.
 
@@ -367,7 +360,7 @@ class MPI:
         data : array
             Distributed data set.
         outlist : bool, optional
-            If outlist is False, we collect and concatenate, if True then 
+            If outlist is False, we collect and concatenate, if True then
             we do not concatenate the list.
         """
         if np.isscalar(data):
@@ -375,18 +368,17 @@ class MPI:
         if self.rank == 0:
             datas = [data]
             for i in range(1, self.size):
-                _data = self.recv(i, tag=10+i)
+                _data = self.recv(i, tag=10 + i)
                 datas.append(_data)
             if outlist is False:
                 data = np.concatenate(datas)
             else:
                 data = datas
         else:
-            self.send(data, to_rank=0, tag=10+self.rank)
+            self.send(data, to_rank=0, tag=10 + self.rank)
             data = None
         self.wait()
         return data
-
 
     def collect_noNone(self, data: np.ndarray) -> np.ndarray:
         """
@@ -407,7 +399,6 @@ class MPI:
             datas = None
         return datas
 
-
     def distribute(self, data: np.ndarray) -> np.ndarray:
         """
         Distribute and split data from rank 0.
@@ -420,15 +411,16 @@ class MPI:
         if self.rank == 0:
             split1, split2 = self.split(len(data))
             for i in range(1, len(split1)):
-                self.send(data[split1[i]:split2[i]], to_rank=i, tag=10+i)
-            data = data[split1[0]:split2[0]]
+                self.send(data[split1[i] : split2[i]], to_rank=i, tag=10 + i)
+            data = data[split1[0] : split2[0]]
         else:
-            data = self.recv(0, tag=10+self.rank)
+            data = self.recv(0, tag=10 + self.rank)
         self.wait()
         return data
 
-
-    def sum(self, data: Union[np.ndarray, int, float]) -> Union[np.ndarray, int, float, None]:
+    def sum(
+        self, data: Union[np.ndarray, int, float]
+    ) -> Union[np.ndarray, int, float, None]:
         """
         Sums a distributed data set to the processor with rank=0.
 
@@ -439,14 +431,13 @@ class MPI:
         """
         if self.rank == 0:
             for i in range(1, self.size):
-                _data = self.recv(i, tag=10+i)
+                _data = self.recv(i, tag=10 + i)
                 data += _data
         else:
-            self.send(data, to_rank=0, tag=10+self.rank)
+            self.send(data, to_rank=0, tag=10 + self.rank)
             data = None
         self.wait()
         return data
-
 
     def mean(self, data: np.ndarray) -> Union[float, int]:
         """
@@ -469,7 +460,6 @@ class MPI:
         self.wait()
         return mean
 
-
     def min(self, data: np.ndarray) -> Union[float, int]:
         """
         Finds the minimum of a distributed data set, which is broadcasted to all nodes.
@@ -488,7 +478,6 @@ class MPI:
         self.wait()
         return minval
 
-
     def max(self, data: np.ndarray) -> Union[float, int]:
         """
         Finds the maximum of a distributed data set, which is broadcasted to all nodes.
@@ -506,7 +495,6 @@ class MPI:
             maxval = self.recv(0, tag=11)
         self.wait()
         return maxval
-
 
     def end(self):  # pragma: no cover
         """
